@@ -1,31 +1,116 @@
 [project]
 name = "agent-memory"
 version = "1.0.0"
-description = "Deterministic, schema-enforced, repo-local Memory Control Plane (MCP) for AI agents."
+description = "A deterministic Memory Control Plane (MCP) server for AI agents with structured, versioned, repository-backed memory system"
+authors = [
+    {name = "MCP Generator", email = "mcp@example.com"}
+]
 readme = "README.md"
-requires-python = ">=3.11"
-license = "MIT"
-license-files = ["LICENSE"]
-authors = [{ name = "Agent Memory MCP", email = "noreply@example.com" }]
+license = {text = "MIT"}
+requires-python = ">=3.13"
+keywords = ["agent", "memory", "mcp", "model-context-protocol", "ai", "persistence"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+]
+dependencies = [
+    # Core MCP server functionality
+    "mcp>=1.0.0",
+    # No additional dependencies needed - uses only Python standard library
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0.0",
+    "pytest-asyncio>=0.21.0",
+    "black>=23.0.0",
+    "isort>=5.12.0",
+    "mypy>=1.0.0",
+]
 
 [project.urls]
-Homepage = "https://example.com/agent-memory"
+Homepage = "https://github.com/example/agent-memory-mcp"
+Repository = "https://github.com/example/agent-memory-mcp"
+Documentation = "https://github.com/example/agent-memory-mcp#readme"
+Issues = "https://github.com/example/agent-memory-mcp/issues"
 
 [project.scripts]
 agent-memory = "agent_memory.__main__:main"
 
-[tool.uv]
-# uv-compatible project; lightweight, no external deps by default
-
 [build-system]
-requires = ["setuptools>=77", "wheel"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-[tool.agent_memory]
-schema_version_default = "v1"
-file_size_max_bytes = 1048576
-directory_listing_limit = 1000
+[tool.hatch.build.targets.wheel]
+# Use src/ layout explicit package path so hatchling includes code.
+packages = ["src/agent_memory"]
 
-[tool.setuptools.packages.find]
-where = ["src"]
-include = ["agent_memory*"]
+[tool.black]
+line-length = 100
+target-version = ['py313']
+include = '\.pyi?$'
+extend-exclude = '''
+/(
+  # directories
+  \.eggs
+  | \.git
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | build
+  | dist
+)/
+'''
+
+[tool.isort]
+profile = "black"
+line_length = 100
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+ensure_newline_before_comments = true
+
+[tool.mypy]
+python_version = "3.13"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = true
+disallow_incomplete_defs = true
+check_untyped_defs = true
+disallow_untyped_decorators = true
+no_implicit_optional = true
+warn_redundant_casts = true
+warn_unused_ignores = true
+warn_no_return = true
+warn_unreachable = true
+strict_equality = true
+
+[[tool.mypy.overrides]]
+module = [
+    "mcp.*",
+]
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py", "*_test.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+addopts = [
+    "-v",
+    "--tb=short",
+    "--strict-markers",
+    "--disable-warnings",
+]
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "integration: marks tests as integration tests",
+]
+asyncio_mode = "auto"
