@@ -1,38 +1,44 @@
 #!/usr/bin/env python3
 """
-Agent Memory MCP Server entrypoint.
+Agent Memory MCP Server Entry Point
 
 Run:
-  uvx python -m agent_memory
+  uvx python -m agent_memory                # start server (stdio, waits for MCP client)
+  uvx python -m agent_memory --test         # run internal self-tests then exit
 """
+
+import argparse
 import asyncio
 import sys
-from typing import NoReturn
+
+from agent_memory.server import run_server, test_server
 
 
-def main() -> int:
-    """Run the MCP server and return POSIX exit code."""
+def parse_args(argv: list[str]) -> argparse.Namespace:
+    parser = argparse.ArgumentParser(prog="agent_memory", add_help=True)
+    parser.add_argument(
+        "--test",
+        action="store_true",
+        help="Run built-in server self tests (tool & resource listing) then exit."
+    )
+    return parser.parse_args(argv)
+
+
+def main():
+    """CLI dispatcher for Agent Memory MCP Server."""
+    args = parse_args(sys.argv[1:])
     try:
-        # Import inside function to reduce import-time side effects for linters
-        from agent_memory.server import run  # type: ignore
-    except ImportError as exc:
-        sys.stderr.write(f"Failed to import server: {exc}\n")
-        return 1
-    try:
-        asyncio.run(run())
-        return 0
+        if args.test:
+            # Run lightweight self-test (does NOT start persistent server loop)
+            asyncio.run(test_server())
+        else:
+            asyncio.run(run_server())
     except KeyboardInterrupt:
-        # Graceful shutdown
-        return 0
-    except Exception as exc:  # pylint: disable=broad-except
-        sys.stderr.write(f"Server runtime error: {exc}\n")
-        return 1
-
-
-def _start() -> NoReturn:
-    """Invoke main() and exit with its return code."""
-    raise SystemExit(main())
+        print("\nServer stopped by user", file=sys.stderr)
+    except Exception as e:
+        print(f"Server error: {e}", file=sys.stderr)
+        sys.exit(1)
 
 
 if __name__ == "__main__":
-    _start()
+    main()
